casewm: Changed to my personal penrose fork.
This is a temporary change, but for now since the main penrose package
has not released a new version, I am using my person fork of penrose
on my local system to build case-wm. This will eventually change as I
plan to move penrose into the case-wm repository and modify penrose to
my liking without depening on upstream.